Adjusted 'amperageLatest' data type to be 'int32_t' to be consistent with 'amperage'.

+++ src/main/sensors/battery.h
@@ -74,7 +74,7 @@ extern uint16_t vbatRaw;
 extern uint16_t vbatLatest;
 extern uint8_t batteryCellCount;
 extern uint16_t batteryWarningVoltage;
-extern uint16_t amperageLatest;
+extern int32_t amperageLatest;
 extern int32_t amperage;
 extern int32_t mAhDrawn;
